Remove references to can/util/jquery
