--- util/jquery/test.html
@@ -1,3 +1,0 @@
-<title>can/util/jquery</title>
-<script src="../../node_modules/steal/steal.js" main="can/util/jquery/jquery_test"></script>
-<div id="qunit-fixture"></div>
