+++ util/test.html
@@ -1,0 +1,3 @@
+<title>can/util/util</title>
+<script src="../../node_modules/steal/steal.js" main="can/util/util_test"></script>
+<div id="qunit-fixture"></div>
