+++ util/util_test.js
@@ -1,7 +1,7 @@
-var can = require('can/util/jquery/jquery');
+var can = require('can/util/util');
 require('steal-qunit');
 
-QUnit.module("can/util/jquery");
+QUnit.module("can/util/util");
 
 // Problem inspired by attempting to bind a compoments templated event listeners to:
 // https://github.com/vote539/socketio-file-upload/blob/master/client.js
